change month and date
